are-they-similar-app/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-vision==2.7.1
google-cloud-storage==2.1.0

